Retrieved data usingHTTP get method from the service
Removed the hard coding in the service and added functionality for displaying of the errors as well.

Files affected include:
app.module.ts - Added the HttpClient Module

product-list.component.ts - Modified the constructor and also included the error message as well. Added the subscription to the pulling.

product.service.ts - Modified so as to have the data not hard coded but read from a JSON file.